De-duplicate validation
